refactor(rpc): Move Emitter to src/emitter.rs
The Emitter struct, its impl, MempoolEvent, BlockEvent, PollResponse,
the poll/poll_once helpers, NO_EXPECTED_MEMPOOL_TXS, and the inline
test module are relocated from lib.rs into a new src/emitter.rs.
lib.rs now declares `mod emitter` and re-exports its public items,
keeping the crate's public API unchanged.
